combine those two change handlers because there's no need to create two change handler for notes. the property is that much as lyric.
